+++ R/DOSPERT_raw_doc.R
@@ -2,7 +2,7 @@
 #'
 #' A data.frame containing responses to the risk subscale of the Domain Specific
 #' Risk Taking Scale (DOSPERT; Weber, Blais, & Betz, 2002) based on the publicly
-#' available dataset (at <https://osf.io/pjt57/>) by Frey, Duncan, and Weber (2020).
+#' available dataset (at <https://osf.io/pjt57/>) by Frey, Duncan, and Weber (2023).
 #' The items measure risk-taking propensity on five different domains: social,
 #' recreational, financial, health/ safety, and ethical.
 #'
@@ -16,6 +16,6 @@
 #'   \item{socR_1 to socR_6}{(numeric) - Social-domain risk-taking items.}
 #'  }
 #' @source Blais, A.-R., & Weber, E. U. (2006). A domain-specific risk-taking (DOSPERT) scale for adult populations. Judgment and Decision Making, 1(1), 33–47. doi: 10.1017/S1930297500000334
-#' @source Frey, R., Duncan, S. M., & Weber, E. U. (2020). Towards a typology of risk preference: Four risk profiles describe two thirds of individuals in a large sample of the U.S. population. PsyArXiv Preprint. doi:10.31234/osf.io/yjwr9
+#' @source Frey, R., Duncan, S. M., & Weber, E. U. (2023). Towards a typology of risk preference: Four risk profiles describe two-thirds of individuals in a large sample of the U.S. population. Journal of Risk and Uncertainty, 66(1), 1–17. doi:10.1007/s11166-022-09398-5
 #'
 "DOSPERT_raw"